Fix multiple device selections (read comment)
Fixed the crash upon entering something into the omnibar that would spawn the device selector.

This required re-writing large, core parts of GAM.cs, so I'll want to keep this in alpha and continue testing before I merge to the normal branch.

The crash was caused by GAM.GetDeviceId opening a UI dialog when it's running in a background thread. To resolve this, I made GAM.GetDeviceId return a List<BasicDeviceInfo>, then I added a function to the BasicDeviceInfo class (BasicDeviceInfo.HandleGetDeviceId) that takes in the list of devices, is called from the main thread, and handles the list, and, if applicable, opening the device selector from the correct (UI) thread this time.

Hopefully that made sense. I also added a new function (GetInput.GetDeviceSelection) which is a wrapper over GetInput.GetDataGridSelection that, when passed in a list of BasicDeviceInfo objects, returns a BasicDeviceInfo object that the user selected, rather than returning a List<string> then parsing that somewhere else.

This was a small commit by lines, but a large amount of logic is here. By the way, to my disbelief, after re-writing GAM.cs, adding methods to BasicDeviceInfo and adding the GetInput.GetDeviceSelection function, it worked perfectly the first time. What are the odds?

I've updated the Software.Version to 1.0.2, and I've updated the installer to match. Most likely, I'll be switching to Nullsoft for the installer to remove the issue of having to uninstall before installing newer versions before merging this branch and releasing 1.0.